+++ src/prefect/schedules/schedules.py
@@ -3,7 +3,7 @@ import itertools
 import operator
 import warnings
 from datetime import datetime, timedelta
-from typing import Callable, Iterable, List, Optional
+from typing import Callable, Iterable, List, Optional, cast
 
 import prefect.schedules.adjustments
 import prefect.schedules.clocks
@@ -211,7 +211,9 @@ def IntervalSchedule(
     return Schedule(
         clocks=[
             prefect.schedules.clocks.IntervalClock(
-                interval=interval, start_date=start_date, end_date=end_date
+                interval=cast(timedelta, interval),  # due to FIXME, above
+                start_date=start_date,
+                end_date=end_date,
             )
         ]
     )
@@ -248,3 +250,31 @@ def CronSchedule(
             )
         ]
     )
+
+
+def OneTimeSchedule(start_date: datetime) -> Schedule:
+    """
+    A schedule corresponding to a single date
+
+    NOTE: this function is deprecated and maintained only for backwards-compatibility.
+    """
+    warnings.warn(
+        "The OneTimeSchedule is deprecated and will be removed from "
+        "Prefect. Use a Schedule with a single-date DatesClock instead.",
+        UserWarning,
+    )
+    return Schedule(clocks=[prefect.schedules.clocks.DatesClock(dates=[start_date])])
+
+
+def UnionSchedule(schedules: List[Schedule]) -> Schedule:
+    """
+    A schedule formed by combining other schedules.
+
+    NOTE: this function is deprecated and maintained only for backwards-compatibility.
+    """
+    warnings.warn(
+        "The UnionSchedule is deprecated and will be removed from "
+        "Prefect. Use a Schedule with multiple clocks instead.",
+        UserWarning,
+    )
+    return Schedule(clocks=[c for s in schedules for c in s.clocks])